refactor: move helper modules to a separated folder
