+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file. Dates are d
 
 - refactor: modularize watch/build process [`96943a5`](https://github.com/henriquehbr/eswatch/commit/96943a53e7788bf4bcdec5ddb67415b103142aa6)
 - refactor: modularize CLI options parsing [`50be3b7`](https://github.com/henriquehbr/eswatch/commit/50be3b7223dfaac175ad7df164c81e4ea32f9de4)
-- refactor: return build steps through Promise.all() [`bcfb547`](https://github.com/henriquehbr/eswatch/commit/bcfb5477e3cda04d1e1357974de3ac60f78d4938)
+- refactor: return build steps through Promise.all() [`f4f9b2b`](https://github.com/henriquehbr/eswatch/commit/f4f9b2b818754f63b38a777e971f1946d804104e)
 - refactor: transform relative imports into absolute [`105b0bb`](https://github.com/henriquehbr/eswatch/commit/105b0bbcc811f6af69c95a5e4af3fb42f4444b60)
+- refactor: move helper modules to a separated folder [`fd035f2`](https://github.com/henriquehbr/eswatch/commit/fd035f28c7ea149e7d6ff937b18033e9bf351554)
 - refactor(parseOptions): add "incremental" option [`3b60fc3`](https://github.com/henriquehbr/eswatch/commit/3b60fc3e94464c452f2e23a2d83cf896e6e7ce8a)
 - chore: bump dependencies [`f183e10`](https://github.com/henriquehbr/eswatch/commit/f183e10b058040f5d777209bc9985bfadbb44386)
