--- lib/getDependencies.js
@@ -1,1 +1,0 @@
-import{readFileSync as o}from"fs";import p from"find-up";const a=async()=>{const e=await p("package.json");if(e){const c=o(e,{encoding:"utf-8"}),n=JSON.parse(c),s=Object.keys(n.dependencies),t=Object.keys(n.devDependencies);return[...s,...t]}return[]};export{a as getDependencies};
